DR2-1329 Resources for e2e-tests
This adds scan on push at the registry level which is the recommended
way of doing it now.
It also adds the task execution role to the repository policy so the
intg task can access the repository in the management account.
I've added ecs:* permissions to terraform so I can deploy the task
definition.