DR2 Move TNA to Preservica role.
To create the DataSync locations and tasks, the role which creates it needs to be able to access the Preservica bucket. This means that terraform needs to be able to assume the role which means it needs to already exist in dr2-terraform-backend.

It’s probably safer to move the resources from the terraform environments state to the terraform backend state rather than delete and recreate in case we break something.

This role is a bit of a weird hybrid but it's working.