Update terraform role.
I had the permissions set on the trust policy for the copy to Preservica
role to allow the GitHub terraform repositories role access which is
wrong.

I've added the environment GitHub roles to the trust policy for each of
the copy to Preservica roles and updated the GitHub roles to assume the
copy to Preservica role.

Terraform is no longer complaining that it can't assume the role.