DR2-1638 Lock down GitHub roles.
A lot of the complexity here comes from the fact that we're using the
same role to run terraform for dr2-terraform-environments,
tdr-aws-accounts and tna-custodian.

Whether this is a good idea is a subject for another time but it's
basically set up like this:

If a role is for a specific environment, e.g. MgmtDPGithubTerraformEnvironmentsRoleIntg
then that role has permissions for the intg environment on GitHub and
for the main branch of the repos that need it.

If the role is for all environments, e.g. MgmtDPGithubCodeDeploy then
this includes all environments.

I've created a module in here to replace duplication. I don't think
we'll need it elsewhere as these are the most complicated permissions
so I've not put it in da-terraform-modules for now.