Update trust policies.
There are a few repos which have tests run on pull_request triggers
which need that explicitly in the trust policy. I'm being slightly lazy
here in that I'm adding it to all repos, not just the ones that need it
but I think that's fine.

I've also removed a couple of old repos that we're no longer using from
the policy.